# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, pending]
        - name: paymentStatus
          in: query
          schema:
            type: string
            enum: [paid, pending, overdue]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Enroll a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payment records
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [completed, pending, failed, refunded]
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    put:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, pending]
        courseId:
          type: integer
          nullable: true
        courseName:
          type: string
          nullable: true
        paymentStatus:
          type: string
          enum: [paid, pending, overdue]
        progress:
          type: number
        enrolledAt:
          type: string
      required: [id, name, email, status, paymentStatus, progress, enrolledAt]
    StudentDetail:
      allOf:
        - $ref: "#/components/schemas/Student"
        - type: object
          properties:
            payments:
              type: array
              items:
                $ref: "#/components/schemas/Payment"
          required: [payments]
    CreateStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        courseId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, inactive, pending]
        paymentStatus:
          type: string
          enum: [paid, pending, overdue]
      required: [name, email, status, paymentStatus]
    Payment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
          nullable: true
        courseId:
          type: integer
          nullable: true
        courseName:
          type: string
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [completed, pending, failed, refunded]
        method:
          type: string
          enum: [bank_transfer, cash, card, online]
        notes:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, studentId, amount, status, method, createdAt]
    CreatePaymentBody:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [completed, pending, failed, refunded]
        method:
          type: string
          enum: [bank_transfer, cash, card, online]
        notes:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
      required: [studentId, amount, status, method]
...