# lib/api-zod/src/generated/types/createStudentBody.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const CreateStudentBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  courseId: zod.number().nullish(),
  status: zod.enum(["active", "inactive", "pending"]),
  paymentStatus: zod.enum(["paid", "pending", "overdue"]),
});

export type CreateStudentBody = zod.input<typeof CreateStudentBody>;