# lib/api-zod/src/generated/types/studentDetail.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const StudentDetail = zod
  .object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    phone: zod.string().nullish(),
    status: zod.enum(["active", "inactive", "pending"]),
    courseId: zod.number().nullish(),
    courseName: zod.string().nullish(),
    paymentStatus: zod.enum(["paid", "pending", "overdue"]),
    progress: zod.number(),
    enrolledAt: zod.string(),
  })
  .and(
    zod.object({
      payments: zod.array(
        zod.object({
          id: zod.number(),
          studentId: zod.number(),
          studentName: zod.string().nullish(),
          courseId: zod.number().nullish(),
          courseName: zod.string().nullish(),
          amount: zod.number(),
          status: zod.enum(["completed", "pending", "failed", "refunded"]),
          method: zod.enum(["bank_transfer", "cash", "card", "online"]),
          notes: zod.string().nullish(),
          paidAt: zod.string().nullish(),
          createdAt: zod.string(),
        }),
      ),
    }),
  );

export type StudentDetail = zod.input<typeof StudentDetail>;